+++ components/Employee/Activator/Nationality.vue
@@ -12,6 +12,7 @@ import { useDefaults } from "vuetify";
  * DEFINE PROPS
  *****************************************************************************/
 const _props = defineProps({
+  canEdit: { type: Boolean, default: false },
   item: {
     type: Object,
     required: true,
@@ -86,7 +87,8 @@ defineExpose({
     <v-toolbar color="secondary" density="compact" :title="props.title">
       <template #append>
         <v-btn
-          icon="mdi-pencil"
+          v-if="props.canEdit"
+            icon="mdi-pencil"
           size="small"
           @click="emit('click:edit', props.item)"
         />