+++ functions/apis/index.js
@@ -28,6 +28,7 @@ export { updateCompanyArrangement } from "./updateCompanyArrangement.js";
 export { updateCompanyBilling } from "./updateCompanyBilling.js";
 export { updateCompanyOperations } from "./updateCompanyOperations.js";
 export { updateCompanyProfile } from "./updateCompanyProfile.js";
+export { createEmployee, updateEmployeeBasic, updateEmployeeNationality } from "./saveEmployee.js";
 export { updateSiteAgreements } from "./updateSiteAgreements.js";
 export {
   updateOwnUserProfile,